src/exchanges/okx/modules/authed/key/fetchDetails.ts
import { debug } from 'debug'

import { IAlunaExchangeAuthed } from '../../../../../lib/core/IAlunaExchange'
import { AlunaHttpVerbEnum } from '../../../../../lib/enums/AlunaHtttpVerbEnum'
import {
  IAlunaKeyFetchDetailsParams,
  IAlunaKeyFetchDetailsReturns,
} from '../../../../../lib/modules/authed/IAlunaKeyModule'
import { OkxHttp } from '../../../OkxHttp'
import { getOkxEndpoints } from '../../../okxSpecs'
import { IOkxKeySchema } from '../../../schemas/IOkxKeySchema'



const log = debug('alunajs:okx/key/fetchDetails')



export const fetchDetails = (exchange: IAlunaExchangeAuthed) => async (
  params: IAlunaKeyFetchDetailsParams = {},
): Promise<IAlunaKeyFetchDetailsReturns> => {

  log('fetching Okx key permissions')

  const {
    settings,
    credentials,
  } = exchange

  const { http = new OkxHttp(settings) } = params

  // TODO: Implement proper request
  const rawKey = await http.authedRequest<IOkxKeySchema>({
    verb: AlunaHttpVerbEnum.GET,
    url: getOkxEndpoints(settings).key.fetchDetails,
    credentials,
  })

  const { key } = exchange.key.parseDetails({ rawKey })

  const { requestWeight } = http

  return {
    key,
    requestWeight,
  }

}
